refactor: moved test in packages
