+++ app/src/test/java/persistence/ebike/AbstractEBikeRepositoryTest.java
@@ -1,6 +1,5 @@
-package persistence;
+package persistence.ebike;
 
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import sap.ass01.layered.persistence.DTO.EBikeDTO;
