+++ app/src/test/java/persistence/ebike/EBikeMySqlTest.java
@@ -1,7 +1,6 @@
-package persistence;
+package persistence.ebike;
 
 import org.junit.jupiter.api.AfterEach;
-import sap.ass01.layered.persistence.MongoDB.MongoEBikeRepository;
 import sap.ass01.layered.persistence.mysql.MySqlEBikeRepository;
 import sap.ass01.layered.persistence.repository.EBikeRepository;
